+++ src/components/general/Button/Button.tsx
@@ -7,30 +7,18 @@ import { ConfigProvider } from 'src/components/other/ConfigProvider/ConfigProvid
 import { type ReactNode } from 'react'
 import './button.css'
 
-export interface IButtonProps extends Omit<AntButtonProps, 'variant' | 'color'> {
-  /**
-   * @deprecated This variant is a temporary fix for new icons.
-   * Use this variant only with new icons to align the icon and text centered.
-   * This will be removed once all icons are updated.
-   */
-  variant?: 'with-new-icon'
+export interface IButtonProps extends AntButtonProps {
   icon?: ReactNode
-  /**
-   * Color of the button text. Use 'inherit' to inherit from parent element.
-   */
-  color?: 'inherit' | string
+  textColor?: 'inherit'
 }
 
-interface ButtonComponent
-  extends React.ForwardRefExoticComponent<IButtonProps & React.RefAttributes<HTMLButtonElement | HTMLAnchorElement>> {
+interface ButtonComponent extends React.ForwardRefExoticComponent<
+  IButtonProps & React.RefAttributes<HTMLButtonElement | HTMLAnchorElement>
+> {
   Group: typeof AntButton.Group
 }
 
 export const Button = React.forwardRef<HTMLButtonElement | HTMLAnchorElement, IButtonProps>((props, ref) => {
-  const classMap = {
-    'with-new-icon': 'u-display-flex u-align-items-center u-justify-center',
-  }
-
   const buttonIcon =
     React.isValidElement<IIconProps>(props.icon) && props.icon.type === Icon ? (
       <Icon {...props.icon.props} color={props.icon.props.color ?? 'inherit'} />
@@ -38,11 +26,9 @@ export const Button = React.forwardRef<HTMLButtonElement | HTMLAnchorElement, IB
       props.icon
     )
 
-  const { variant, color, style, ...restProps } = props
+  const { variant, textColor, style, ...restProps } = props
 
-  const extraClass = variant === 'with-new-icon' ? classMap['with-new-icon'] : ''
-
-  const buttonStyle = color ? { ...style, color } : style
+  const buttonStyle = textColor === 'inherit' ? { ...style, color: 'inherit' } : style
 
   return (
     <ConfigProvider>
@@ -51,7 +37,8 @@ export const Button = React.forwardRef<HTMLButtonElement | HTMLAnchorElement, IB
         ref={ref}
         icon={buttonIcon}
         style={buttonStyle}
-        className={`${props.className ?? ''} ${extraClass}`}>
+        variant={variant}
+        className={`${props.className ?? ''}`}>
         {props.children}
       </AntButton>
     </ConfigProvider>