+++ src/components/index.ts
@@ -1,3 +1,6 @@
+import '@fontsource/inter'
+import '@fontsource/courier-prime'
+
 export { Button, type IButtonProps } from './general/Button/Button'
 export { FloatButton, type IFloatButtonProps } from './general/FloatButton/FloatButton'
 export { Icon, type IIconProps } from './general/Icon/Icon'
@@ -162,3 +165,53 @@ export {
   UnauthorizedTooltip,
   type IUnauthorizedTooltipProps,
 } from './UXPatterns/PermissionsRestrictions/UnauthorizedTooltip'
+
+// Export Rokt icons from @untitledui/icons
+export {
+  RoktHelpCircle,
+  RoktPlus,
+  RoktSearch,
+  RoktCheckCircle,
+  RoktArrowLeft,
+  RoktArrowRight,
+  RoktTag,
+  RoktAlertCircle,
+  RoktAnnouncement,
+  RoktArrowDown,
+  RoktAsterisk,
+  RoktCheck,
+  RoktChevronDown,
+  RoktChevronLeft,
+  RoktChevronRight,
+  RoktChevronSelector,
+  RoktClock,
+  RoktCoinsHand,
+  RoktCopy,
+  RoktDotsVertical,
+  RoktEdit,
+  RoktEye,
+  RoktInfoCircle,
+  RoktLinkExternal,
+  RoktLightbulb,
+  RoktMail,
+  RoktMonitor,
+  RoktPauseCircle,
+  RoktPhoneIncoming,
+  RoktRefreshCw,
+  RoktSearchLg,
+  RoktStars,
+  RoktTrash,
+  RoktUser,
+  RoktX,
+  RoktXClose,
+  RoktBuilding,
+  RoktGlobe,
+  RoktLayoutAlt,
+  RoktShoppingCart,
+  RoktMarkerPin,
+  RoktZoomIn,
+  RoktImage,
+  RoktMessageChatSquare,
+  RoktUserCheck,
+  RoktClipboard,
+} from './icons'